Resolve "Create service for Order"

See merge request xbaran4/bookhub!47

+++ PV179_BookHub/BusinessLayer/Facades/Publisher/PublisherFacade.cs
@@ -1,22 +1,23 @@
 ﻿using AutoMapper;
 using BusinessLayer.DTOs.Publisher.Create;
 using BusinessLayer.DTOs.Publisher.View;
+using BusinessLayer.Exceptions;
 using BusinessLayer.Services;
 
 namespace BusinessLayer.Facades.Publisher
 {
     public class PublisherFacade : BaseFacade, IPublisherFacade
     {
-        private readonly IGenericService<DataAccessLayer.Models.Publication.Publisher, long> _publisherService;
+        private readonly IGenericService<PublisherEntity, long> _publisherService;
 
-        public PublisherFacade(IMapper mapper, IGenericService<DataAccessLayer.Models.Publication.Publisher, long> publisherService) : base(mapper)
+        public PublisherFacade(IMapper mapper, IGenericService<PublisherEntity, long> publisherService) : base(mapper)
         {
             _publisherService = publisherService;
         }
 
         public async Task<DetailedPublisherViewDto> CreatePublisherAsync(CreatePublisherDto createPublisherDto)
         {
-            var publisher = _mapper.Map<DataAccessLayer.Models.Publication.Publisher>(createPublisherDto);
+            var publisher = _mapper.Map<PublisherEntity>(createPublisherDto);
             await _publisherService.CreateAsync(publisher);
 
             return _mapper.Map<DetailedPublisherViewDto>(publisher);
@@ -25,6 +26,10 @@ namespace BusinessLayer.Facades.Publisher
         public async Task DeletePublisherAsync(long id)
         {
             var publisher = await _publisherService.FindByIdAsync(id);
+            if (publisher.Books != null && publisher.Books.Any())
+            {
+                throw new RemoveErrorException(typeof(PublisherEntity), typeof(BookEntity));
+            }
 
             await _publisherService.DeleteAsync(publisher);
         }
